requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
playwright==1.32.1
pluggy==1.0.0
pyee==9.0.4
pytest==7.3.1
pytest-base-url==2.0.0
pytest-playwright==0.3.3
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.29.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
